requirements.txt
python-telegram-bot==20.3
pymongo
requests
dotenv
dnspython